Revert timeout version of script to previous state
Turns out this strategy doesn't work in the end. Oddly enough it worked
for a simple test of using `grep` and `grep --help` to test it but it
did not work with the PostgreSQL example from the docs.

It always caused a timeout and it also ate all of the stdout / stderr in
that case.